Services/ContentderAI.SimilarWord/requirements.txt
Django==3.1
gensim==3.8.3
djangorestframework==3.11.1
python-decouple==3.3
gunicorn==20.0.4
whitenoise==5.0.1
